fix sign conversion and fread incorrect handling

+++ source/decoders/palette.cpp
@@ -10,12 +10,12 @@ void Palette::Decode(const char* filename)
     FILE* file = fopen(filename, "rb");
     if (file) {
         colors.resize(256);
-        for (int i = 0; i < 256; i++)
+        for (size_t i = 0; i < 256; i++)
         {
             // order is BGR, not RGB
-            colors[i].b = (uint8_t)fgetc(file);
-            colors[i].g = (uint8_t)fgetc(file);
-            colors[i].r = (uint8_t)fgetc(file);
+            colors[i].b = static_cast<uint8_t>(fgetc(file));
+            colors[i].g = static_cast<uint8_t>(fgetc(file));
+            colors[i].r = static_cast<uint8_t>(fgetc(file));
         }
         fclose(file);
         _isValid = true;
